Prototype-2/requirements.txt
python-json-logger>=2.0.7
pydantic>=2.10.6
pdfplumber>=0.11.5
chonkie[openai]>=0.4.2
openai>=1.68.2
python-dotenv>=1.0.1
qdrant-client>=1.13.3
langdetect>=1.0.9
tiktoken>=0.9.0
